Created LoanController and RefundController plus changes to existing files

+++ src/main/java/com/car_rental/car_rental/entites/Refund.java
@@ -19,6 +19,6 @@ public class Refund {
     @OneToOne
     @JoinColumn(name = "reservation_id")
     private Reservation reservation;
-    private Long surcharge;
+    private Double surcharge;
     private String comment;
 }
